+++ src/integrations/skadi/api/v1/types.ts
@@ -1,23 +1,10 @@
 import type { User } from '../../../../entity';
-
-export interface EstimatedReach {
-  impressions: number;
-  clicks: number;
-  users: number;
-  min_impressions: number;
-  max_impressions: number;
-}
+import type { CancelCampaignArgs, EstimatedReachResponse } from '../common';
 
 export type LegacyPostEstimatedReach = Pick<
   EstimatedReachResponse,
   'clicks' | 'impressions' | 'users'
 >;
-
-export interface EstimatedReachResponse
-  extends Pick<EstimatedReach, 'impressions' | 'clicks' | 'users'> {
-  minImpressions: number;
-  maxImpressions: number;
-}
 
 export interface PromotedPost {
   campaign_id: string;
@@ -77,11 +64,6 @@ export interface StartPostCampaignParams {
   userId: string;
   budget: number;
   durationInDays: number;
-}
-
-export interface CancelCampaignArgs {
-  campaignId: string;
-  userId: string;
 }
 
 export interface ISkadiApiClientV1 {
