+++ src/integrations/skadi/index.ts
@@ -1,3 +1,5 @@
 export * from './types';
 export { SkadiClient, skadiPersonalizedDigestClient } from './clients';
 export * from './api/v1';
+export * from './api/v2';
+export * from './api/common';
